requirements.txt
slack-bolt>=1.18.0
anthropic>=0.39.0
python-docx>=1.1.0
python-dotenv>=1.0.0
